Create a router group (app) and move all frontend pages there.
