+++ src/app/(app)/industry-use-cases/accordion-item.tsx
