+++ src/app/(app)/industry-use-cases/chart-data.ts
