+++ src/app/(app)/industry-use-cases/data.tsx
