+++ src/app/(app)/industry-use-cases/layout.tsx
