+++ src/app/(app)/industry-use-cases/opengraph-image.png
