+++ src/app/(app)/industry-use-cases/page.tsx
