+++ src/app/(app)/industry-use-cases/tab-dependencies.tsx
