+++ src/app/(app)/industry-use-cases/tab-ecosystem.tsx
