+++ src/app/(app)/industry-use-cases/tab-impacts.tsx
