+++ src/app/(app)/industry-use-cases/twitter-image.png
