+++ src/app/(app)/industry-use-cases/types.d.ts
