+++ src/app/(app)/key-concepts/natural-capital-chart/chart-hooks.ts
