+++ src/app/(app)/key-concepts/natural-capital-chart/data.ts
