+++ src/app/(app)/key-concepts/natural-capital-chart/desktop-chart.tsx
