+++ src/app/(app)/key-concepts/natural-capital-chart/index.tsx
