+++ src/app/(app)/key-concepts/natural-capital-chart/mobile-chart.tsx
