+++ src/app/(app)/natural-capital-in-daily-life/data.tsx
