+++ src/app/(app)/natural-capital-in-daily-life/intro.tsx
