+++ src/app/(app)/natural-capital-in-daily-life/layout.tsx
