+++ src/app/(app)/natural-capital-in-daily-life/outro.tsx
