+++ src/app/(app)/natural-capital-in-daily-life/page.tsx
