+++ src/app/(app)/natural-capital-in-daily-life/section.tsx
