+++ src/app/(app)/resources/layout.tsx
