+++ src/app/(app)/resources/page.tsx
