+++ src/app/(app)/terms-and-conditions/layout.tsx
