+++ src/app/(app)/terms-and-conditions/page.tsx
