+++ tsconfig.json
@@ -24,7 +24,7 @@
     "paths": {
       "@/*": [
         "./src/*"
-      ]
+      ],
     },
     "target": "ES2017"
   },
